Added functionality to checkboxes
