+++ src/style.css
@@ -41,6 +41,11 @@ button {
   font-size: 1rem;
 }
 
+.checkedDisplay{
+  text-decoration: line-through;
+  color: rgba(0, 0, 0, 0.37);
+}
+
 .enterBtn {
   padding: 0.7rem;
 }
